First attempt at interpolated infinite phase screen
